+++ go.mod
@@ -5,6 +5,7 @@ go 1.16
 require (
 	github.com/pkg/errors v0.9.1
 	github.com/sirupsen/logrus v1.7.0
+	github.com/spf13/cobra v1.7.0
 	github.com/stretchr/testify v1.6.1
 	go.uber.org/goleak v1.1.10
 )
